Move Symtab to rsc.symtab
