+++ rsc/src/main/scala/rsc/Compiler.scala
@@ -15,6 +15,7 @@ import rsc.report._
 import rsc.scan._
 import rsc.semantics._
 import rsc.settings._
+import rsc.symtab._
 import rsc.syntax._
 import rsc.util._
