+++ rsc/src/main/scala/rsc/outline/Outliner.scala
@@ -7,6 +7,7 @@ import rsc.input._
 import rsc.report._
 import rsc.semantics._
 import rsc.settings._
+import rsc.symtab._
 import rsc.syntax._
 import rsc.util._
 import scala.collection.mutable
