+++ rsc/src/main/scala/rsc/outline/Scheduler.scala
@@ -8,6 +8,7 @@ import rsc.input._
 import rsc.report._
 import rsc.semantics._
 import rsc.settings._
+import rsc.symtab._
 import rsc.syntax._
 import rsc.util._
