+++ rsc/src/main/scala/rsc/outline/Synthesizer.scala
@@ -6,6 +6,7 @@ import rsc.gensym._
 import rsc.report._
 import rsc.semantics._
 import rsc.settings._
+import rsc.symtab._
 import rsc.syntax._
 import rsc.util._
