+++ rsc/src/main/scala/rsc/symtab/Symtab.scala
@@ -1,10 +1,11 @@
 // Copyright (c) 2017-2019 Twitter, Inc.
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
-package rsc.outline
+package rsc.symtab
 
 import java.util.{HashMap, HashSet, LinkedHashMap}
 import rsc.classpath._
 import rsc.input._
+import rsc.outline._
 import rsc.pretty._
 import rsc.semantics._
 import rsc.settings._
